clean up existing iterables
