+++ packages/slate/test/models/node/get-descendants-at-range/multiple-blocks.js
@@ -30,7 +30,7 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
