+++ packages/slate/test/models/node/get-descendants-at-range/nested-blocks-cursor-in-first-leaf-of-first-parent.js
@@ -23,7 +23,7 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
