+++ packages/slate/test/models/node/get-descendants-at-range/nested-blocks-cursor-in-first-leaf-of-second-parent.js
@@ -8,24 +8,30 @@ export const input = (
       <quote key="a">
         <quote key="b">
           <paragraph key="c">
-            <text key="d">
-              <anchor />one<focus />
+            <text key="d">one</text>
+          </paragraph>
+          <paragraph key="e">
+            <text key="f">two</text>
+          </paragraph>
+        </quote>
+        <quote key="g">
+          <paragraph key="h">
+            <text key="i">
+              three
+              <cursor />
             </text>
           </paragraph>
         </quote>
       </quote>
-      <paragraph key="e">
-        <text key="f">two</text>
-      </paragraph>
     </document>
   </value>
 )
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
 
-export const output = ['a', 'b', 'c', 'd']
+export const output = ['a', 'g', 'h', 'i']
