+++ packages/slate/test/models/node/get-descendants-at-range/nested-blocks-cursor-in-second-leaf-of-first-parent.js
@@ -27,7 +27,7 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
