+++ packages/slate/test/models/node/get-descendants-at-range/nested-blocks-selection-overlapping-texts-in-second-parent.js
@@ -5,42 +5,40 @@ import h from '../../../helpers/h'
 export const input = (
   <value>
     <document>
-      <paragraph key="a">
-        <text key="b">
-          <focus />
-          one
-        </text>
-      </paragraph>
-      <quote key="c">
-        <quote key="d">
-          <paragraph key="e">
-            <text key="f">two</text>
+      <quote key="a">
+        <quote key="b">
+          <paragraph key="c">
+            <text key="d">one</text>
           </paragraph>
         </quote>
-        <quote key="g">
+        <quote key="e">
+          <paragraph key="f">
+            <text key="g">
+              <anchor />two
+            </text>
+          </paragraph>
           <paragraph key="h">
             <text key="i">three</text>
           </paragraph>
           <paragraph key="j">
             <text key="k">
-              <anchor />
-              four
+              f<focus />our
             </text>
-          </paragraph>
-          <paragraph key="l">
-            <text key="m">five</text>
           </paragraph>
         </quote>
       </quote>
+      <paragraph key="l">
+        <text key="m">five</text>
+      </paragraph>
     </document>
   </value>
 )
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
 
-export const output = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k']
+export const output = ['a', 'e', 'f', 'g', 'h', 'i', 'j', 'k']
