+++ packages/slate/test/models/node/get-descendants-at-range/single-block-cursor-beginning-of-text.js
@@ -20,7 +20,7 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
