+++ packages/slate/test/models/node/get-descendants-at-range/single-block-cursor-end-of-text.js
@@ -7,11 +7,15 @@ export const input = (
     <document>
       <paragraph key="a">
         <text key="b">one</text>
-        <inline type="link" key="c">
-          <text key="d">
-            tw<cursor />o
-          </text>
-        </inline>
+      </paragraph>
+      <paragraph key="c">
+        <text key="d">
+          two
+          <cursor />
+        </text>
+      </paragraph>
+      <paragraph key="e">
+        <text key="f">three</text>
       </paragraph>
     </document>
   </value>
@@ -19,9 +23,9 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
 
-export const output = ['a', 'c', 'd']
+export const output = ['c', 'd']
