+++ packages/slate/test/models/node/get-descendants-at-range/single-block-cursor-middle-of-text.js
@@ -19,7 +19,7 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
