+++ packages/slate/test/models/node/get-descendants-at-range/single-void-block.js
@@ -18,7 +18,7 @@ export const input = (
 
 export default function({ document, selection }) {
   return document
-    .getNodesAtRange(selection)
+    .getDescendantsAtRange(selection)
     .map(n => n.key)
     .toArray()
 }
